Added a Wait command, which takes and integer. Switched wait parser to  use Wait command, not drive command at 0 speed.

+++ src/org/usfirst/frc/team5687/robot/commands/AutonomousScript.java
@@ -242,7 +242,7 @@ public class AutonomousScript extends CommandGroup {
 			return LogError("Only units milliseconds/seconds accepted in wait command" + tokens[tokenCheck+1]);
 		}
 
-		return new AutonomousWait(waitTime);
+		return new AutonomousWait((int) waitTime);
 
 	}
